==> 00 - wines_and_ratings.ipynb <==
# Initial Operations

## Importing Data

import pandas as pd
import numpy as np

wines = pd.read_csv("wines.csv")
wines

wines.info()

ratings = pd.read_csv("ratings.csv", low_memory=False)
ratings

ratings.info()

## Joining Data

wines_and_ratings = pd.merge(ratings, wines, on='WineID')
wines_and_ratings = wines_and_ratings[["RatingID", "UserID", "Rating", "Vintage", "WineName", "Type", "Elaborate",
                                       "Grapes", "Harmonize", "ABV", "Body", "Acidity", "Country", "RegionName",
                                       "WineryName", "Website"]]
wines_and_ratings

wines_and_ratings.info()

# Preprocessing

## Replacing Some of Special Characters in Specified Columns

wines_and_ratings["Body"] = wines_and_ratings["Body"].str.replace("-", " ")
wines_and_ratings["Body"].head(5)

wines_and_ratings["Grapes"] = wines_and_ratings["Grapes"].str.replace("[", " ").str.replace("]", " ").str.replace("'", "").str.strip()
wines_and_ratings["Grapes"].head(5)

wines_and_ratings["Harmonize"] = wines_and_ratings["Harmonize"].str.replace("[", " ").str.replace("]", " ").str.replace("'", "").str.strip()
wines_and_ratings["Harmonize"].head(5)

wines_and_ratings

## Splitting Specified Columns by a Delimiter and Expanding Them into Multiple Columns

def split_and_expand_columns(df, column_name, delimiter="/"):
    expanded_columns = df[column_name].str.split(delimiter, expand=True)
    new_columns = [f"{column_name}{i+1}" for i in range(expanded_columns.shape[1])]
    expanded_columns.columns = new_columns

    df = pd.concat([df, expanded_columns], axis=1)
    df = df.fillna("")
    df = df.drop(columns=[column_name])

    return df

wines_and_ratings = split_and_expand_columns(wines_and_ratings, "Type")

wines_and_ratings = split_and_expand_columns(wines_and_ratings, "Elaborate")

wines_and_ratings = split_and_expand_columns(wines_and_ratings, "Grapes", delimiter=", ")

wines_and_ratings = split_and_expand_columns(wines_and_ratings, "Harmonize", delimiter=", ")

wines_and_ratings.info()

## Transforming Unreasonable Values in the ABV Column

def transform_values(df, column_name):
    df[column_name] = np.where(df[column_name] > 14.5, 14.5, df[column_name])
    df[column_name] = np.where(df[column_name] < 9, 9, df[column_name])
    return df

wines_and_ratings = transform_values(wines_and_ratings, 'ABV')
wines_and_ratings['ABV'].sort_values()

## Checking and Replacing Special Characters in Some Other Columns

def check_special_characters(df, column_name, special_characters_pattern):
    rows_with_special_characters = df[df[column_name].str.contains(special_characters_pattern, regex=True, na=False)]
    row_count = len(rows_with_special_characters)
    print("Number of rows with special characters in {}: {}".format(column_name, row_count))
    print(rows_with_special_characters[[column_name]].drop_duplicates().sort_values(by=column_name))

import re
def remove_special_characters(df, column_name, characters_to_remove):
    df[column_name] = df[column_name].str.replace(f'[{re.escape(characters_to_remove)}]', ' ', regex=True)
    return df

check_special_characters(wines_and_ratings, 'WineName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

characters_to_remove = r'!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~'
wines_and_ratings = remove_special_characters(wines_and_ratings, 'WineName', characters_to_remove)

check_special_characters(wines_and_ratings, 'WineName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

check_special_characters(wines_and_ratings, 'RegionName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

characters_to_remove = r'!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~'
wines_and_ratings = remove_special_characters(wines_and_ratings, 'RegionName', characters_to_remove)

check_special_characters(wines_and_ratings, 'RegionName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

check_special_characters(wines_and_ratings, 'WineryName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

characters_to_remove = r'!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~'
wines_and_ratings = remove_special_characters(wines_and_ratings, 'WineryName', characters_to_remove)

check_special_characters(wines_and_ratings, 'WineryName', '[!"#$%°&\'()*+,-./:;<=>?@[\\]^_`{|}~]')

## Handling Null and Irregular Values in Website Column

wines_and_ratings['Website'].sample(5)

wines_and_ratings[~wines_and_ratings['Website'].str.startswith("https")][['Website']].sample(5)

patterns_to_remove = ['http://www.', 'https://www.', 'http://', 'https://', 'www.']
pattern = '|'.join(map(re.escape, patterns_to_remove))
wines_and_ratings['Website'] = wines_and_ratings['Website'].str.replace(pattern, '', regex=True)

wines_and_ratings['Website'].sample(5)

wines_and_ratings['Website'] = wines_and_ratings['Website'].replace('', 'vivino.com')

wines_and_ratings['Website'][50289]

wines_and_ratings['Website'][69422]

wines_and_ratings['Website'][132830]

wines_and_ratings['Website'][109011]

# Exporting Data

wines_and_ratings = wines_and_ratings.sort_values(by='RatingID')
wines_and_ratings = wines_and_ratings.reset_index(drop=True)
wines_and_ratings

wines_and_ratings.to_csv('wines_and_ratings.csv', index=False, compression="zstd")
