==> 01 - content_based.ipynb <==
# Modeling Content-Based Filtering Recommender System

## Importing Data

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel
import joblib

df = pd.read_csv('wines_and_ratings.csv', low_memory=False, compression="zstd")
df

df.info()

## Combining Selected Columns into a Single 'Attributes' Column

df = df.fillna("")
common_columns = ['WineName', 'Body', 'Acidity', 'Country', 'RegionName', 'WineryName']
type_columns = ['Type1', 'Type2']
elaborate_columns = ['Elaborate1', 'Elaborate2']
grapes_columns = [f'Grapes{i}' for i in range(1, 10)]
harmonize_columns = [f'Harmonize{i}' for i in range(1, 13)]
all_columns = common_columns + type_columns + elaborate_columns + grapes_columns + harmonize_columns
df['Attributes'] = df[all_columns].astype(str).agg(' '.join, axis=1)
df

## Removing Duplicate Rows based on the 'Attributes' and 'WineName' Columns

df.drop_duplicates(subset=['Attributes'], inplace=True)
df.drop_duplicates(subset=['WineName'], inplace=True)
df.reset_index(drop=True, inplace=True)
df

## Converting 'Attributes' into a TF-IDF Matrix

tfidf = TfidfVectorizer()
tfidf_matrix = tfidf.fit_transform(df['Attributes'])
tfidf_matrix.shape

## Computing Cosine Similarity between Wines based on TF-IDF Matrix

cosine_sim = linear_kernel(tfidf_matrix, tfidf_matrix)

## Testing the Model

def get_recommendations(WineName):
    idx = df.index[df['WineName'] == WineName].tolist()[0]
    sim_scores = list(enumerate(cosine_sim[idx]))
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    sim_scores = sim_scores[1:11]
    wine_indices = [i[0] for i in sim_scores]
    return df['WineName'].iloc[wine_indices]

get_recommendations('Vecchia Grappa Moscato')

## Exporting the Model

joblib.dump((cosine_sim, df), 'content_based_recommender_model.pkl')
