==> 02 - content_based_test.ipynb <==
# Testing Content-Based Filtering Recommender System

import pandas as pd
import joblib
from IPython.display import HTML

cosine_sim_loaded, df_loaded = joblib.load('content_based_recommender_model.pkl')

def get_recommendations(WineName):
    idx = df_loaded.index[df_loaded['WineName'] == WineName].tolist()[0]
    sim_scores = list(enumerate(cosine_sim_loaded[idx]))
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    sim_scores = sim_scores[1:11]
    wine_indices = [i[0] for i in sim_scores]
    
    recommendations = df_loaded.loc[wine_indices, ['WineName', 'Website']].copy()
    recommendations['Website'] = recommendations['Website'].apply(lambda x: f'<a href="{x}" target="_blank">{x}</a>')
    
    display(HTML(recommendations.to_html(escape=False, index=False)))

get_recommendations('Vecchia Grappa Moscato')
