==> 03 - collaborative.ipynb <==
# Modeling Collaborative Filtering Recommender System

## Importing Data

import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity, euclidean_distances
import joblib
from IPython.display import HTML

df_2 = pd.read_csv('wines_and_ratings.csv', low_memory=False, compression="zstd")
df_2

## Creating a Pivot Table to Convert User Ratings to a Wide Format

df_2_wide = pd.pivot_table(df_2, values=['Rating'],
                            index=['WineName', 'UserID'],
                            aggfunc=np.mean).unstack()
df_2_wide = df_2_wide.fillna(0)
df_2_wide

## Calculating Cosine Similarity and Euclidean Distance Matrices

cosine_sim2 = cosine_similarity(df_2_wide)
cosine_sim2_df = pd.DataFrame(cosine_sim2, index=df_2_wide.index, columns=df_2_wide.index)
cosine_sim2_df

dist_euc = euclidean_distances(df_2_wide)
dist_euc_df = pd.DataFrame(dist_euc, index=df_2_wide.index, columns=df_2_wide.index)
dist_euc_df

## Testing the Model

def find_similar_wines_combined(wines, count=10):
    wines_cosine = [wine for wine in wines if wine in cosine_sim2_df.columns]
    wines_summed_cosine = cosine_sim2_df[wines_cosine].apply(lambda row: np.sum(row), axis=1)
    wines_summed_cosine = wines_summed_cosine.sort_values(ascending=False)
    ranked_wines_cosine = wines_summed_cosine.index[wines_summed_cosine.index.isin(wines_cosine) == False]
    ranked_wines_cosine = ranked_wines_cosine.tolist()
    ranked_wines_cosine = ranked_wines_cosine[:count]

    wines_euclidean = [wine for wine in wines if wine in dist_euc_df.columns]
    wines_summed_euclidean = dist_euc_df[wines_euclidean].apply(lambda row: np.sum(row), axis=1)
    wines_summed_euclidean = wines_summed_euclidean.sort_values()
    ranked_wines_euclidean = wines_summed_euclidean.index[wines_summed_euclidean.index.isin(wines_euclidean) == False]
    ranked_wines_euclidean = ranked_wines_euclidean.tolist()
    ranked_wines_euclidean = ranked_wines_euclidean[:count]

    combined_ranking = {}
    for i, wine in enumerate(ranked_wines_cosine):
        combined_ranking[wine] = combined_ranking.get(wine, 0) + i

    for i, wine in enumerate(ranked_wines_euclidean):
        combined_ranking[wine] = combined_ranking.get(wine, 0) + i

    sorted_combined_ranking = sorted(combined_ranking.items(), key=lambda x: x[1])

    top_10_wine_indices = [df_2.index[df_2['WineName'] == wine].tolist()[0] for wine, _ in sorted_combined_ranking[:count]]
    top_10_wines = df_2.loc[top_10_wine_indices, ['WineName', 'Website']].copy()

    top_10_wines['Website'] = top_10_wines['Website'].apply(lambda x: f'<a href="{x}" target="_blank">{x}</a>')

    display(HTML(top_10_wines.to_html(escape=False, index=False)))

find_similar_wines_combined(['Vecchia Grappa Moscato'], 10)

## Exporting the Model

joblib.dump((cosine_sim2_df, dist_euc_df, df_2), 'collaborative_recommender_model.pkl', compress=('zlib', 3))
